+++ codex-rs/otel/src/events/mod.rs
@@ -1,2 +1,2 @@
-pub(crate) mod session;
+pub(crate) mod otel_manager;
 pub(crate) mod shared;
