--- codex-rs/otel/src/events/shared/log.rs
@@ -1,21 +1,0 @@
-macro_rules! log_event {
-    ($self:expr, $($fields:tt)*) => {{
-        tracing::event!(
-            target: $crate::targets::OTEL_LOG_ONLY_TARGET,
-            tracing::Level::INFO,
-            $($fields)*
-            event.timestamp = %$crate::events::shared::timestamp(),
-            conversation.id = %$self.metadata.conversation_id,
-            app.version = %$self.metadata.app_version,
-            auth_mode = $self.metadata.auth_mode,
-            originator = %$self.metadata.originator,
-            user.account_id = $self.metadata.account_id,
-            user.email = $self.metadata.account_email,
-            terminal.type = %$self.metadata.terminal_type,
-            model = %$self.metadata.model,
-            slug = %$self.metadata.slug,
-        );
-    }};
-}
-
-pub(crate) use log_event;
